blackening code
