+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-files: 'hmrm\/'
+files: 'hmrm/'
 repos:
   - repo: https://github.com/psf/black
     rev: 23.3.0
