+++ hmrm/foundation/abs_classes/__pycache__/singleton.cpython-310.pyc
