+++ hmrm/foundation/abs_classes/__pycache__/singleton.cpython-38.pyc
