+++ hmrm/foundation/abs_classes/singleton.py
@@ -1,5 +1,6 @@
 from abc import ABC
 
+
 class Singleton(ABC):
     _instance = None
