+++ hmrm/foundation/configuration/__pycache__/input.cpython-310.pyc
