+++ hmrm/foundation/configuration/__pycache__/input.cpython-38.pyc
